Moving latest Pods bits to Services Plugin
