+++ src/js/stores/__tests__/MesosStateStore-test.js
@@ -1,6 +1,6 @@
 jest.dontMock('../MesosStateStore');
 
-const MesosStateUtil = require('../../../../plugins/services/src/js/utils/MesosStateUtil');
+const MesosStateUtil = require('../../utils/MesosStateUtil');
 const Pod = require('../../../../plugins/services/src/js/structs/Pod');
 const Framework = require('../../../../plugins/services/src/js/structs/Framework');
 const MesosStateStore = require('../MesosStateStore');
